app/frontend/requirements.txt
streamlit
requests
pandas
matplotlib
python-dotenv
plotly 
seaborn
